+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0 beta 4
 * Create a default config file if it does not exist
 * Set colors on tooltips too in the default theme
+* Add customizable statusbar
 
 # 1.0 beta 3
 * Updated README